backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
langgraph
langchain
langchain-core
langchain-openai
langchain-community
scikit-learn
numpy
openai
python-dotenv
matplotlib
seaborn
chromadb
tiktoken
